refactor(commands): move saveRecordingAndTranscribe function to commands.ts and remove handleRecordingTranscription module
